requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
pycparser==2.22
cryptography==44.0.0

charset-normalizer==3.4.0
idna==3.10
urllib3==2.2.3
requests==2.32.3
requests-oauthlib==2.0.0
oauthlib==3.2.2
django-cors-headers

pillow==11.0.0
tzdata==2024.2
sqlparse==0.5.2

Django==5.1.4
django-allauth==65.3.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1

PyJWT==2.10.1
python3-openid==3.2.0
social-auth-core==4.5.4
social-auth-app-django==5.4.2

psycopg2-binary==2.9.10
defusedxml==0.8.0rc2
python-dotenv
